Changement clé primaire entreprise

+++ Handlers/DeleteEntrepriseCommand.cs
@@ -20,11 +20,13 @@ namespace Taxes.Handlers
         }
         public Task<bool> Handle(DeleteEntrepriseCommand request, CancellationToken cancellationToken)
         {
-            Entreprise entreprise = _context.entreprises.FirstOrDefault(ent => ent.Matricule_ciger == request.Matricule);
+            Entreprise entreprise = _context.entreprises.FirstOrDefault(ent => ent.Id_entreprise == request.ID);
             if(entreprise == null)
             {
                 throw new Exception("L'entreprise est introuvable");
             }
+            if (entreprise.Proces_verbal == true) throw new Exception("L'entreprise ne peut pas être supprimée car elle est en infraction");
+
                 _context.entreprises.Remove(entreprise);
                 _context.SaveChanges();
                 return Task.FromResult(true);
